enhance: refact files
