+++ back/src/db/db.manager.js
@@ -1,18 +1,7 @@
-import { db } from './config.js';
+import { LogManager } from './log.manager.js';
+import { MessageManager } from './message.manager.js';
 
 export class DatabaseManager {
-  static saveLog({ createdAt, route, method, input, error, status }) {
-    const insert = db.prepare(
-      `
-      INSERT INTO logs (created_at, route, method, input, error, status)
-      VALUES (?, ?, ?, ?, ?, ?)
-    `,
-    );
-
-    return insert.run(createdAt ?? new Date().toISOString(), route, method, input, error, status);
-  }
-
-  static findAllLogs() {
-    return db.prepare('SELECT * FROM logs ORDER BY id DESC').all();
-  }
+  static log = LogManager;
+  static message = MessageManager;
 }
