+++ back/src/db/log.manager.js
@@ -1,0 +1,18 @@
+import { db } from './config.js';
+
+export class LogManager {
+  static save({ createdAt, route, method, input, error, status }) {
+    const insert = db.prepare(
+      `
+      INSERT INTO logs (created_at, route, method, input, error, status)
+      VALUES (?, ?, ?, ?, ?, ?)
+    `,
+    );
+
+    return insert.run(createdAt ?? new Date().toISOString(), route, method, input, error, status);
+  }
+
+  static findAll() {
+    return db.prepare('SELECT * FROM logs ORDER BY id DESC').all();
+  }
+}
