+++ back/src/db/message.manager.js
@@ -1,0 +1,3 @@
+export class MessageManager {
+  static save() {}
+}
